requirements.txt
# Core Dependencies
streamlit>=1.28.0
flask>=3.0.0
flask-cors>=4.0.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
python-dotenv>=1.0.0

# Database
psycopg2-binary>=2.9.9
sqlparse>=0.4.4

# LLM & LangChain
langchain>=0.1.0
langchain-google-genai>=0.0.11
langchain-groq>=0.0.1
google-generativeai>=0.3.0

# Utilities
Faker>=22.0.0

# Optional (for development)
# jupyter>=1.0.0
# notebook>=7.0.0
